fix: Remove three unbounded memory accesses
Issue #2 asks for the correction of unbounded memory accesses. A search with
AddressSanitizer and with builds at the limits found three more.

1. The binary search gave an index equal to the size of the list.

The search stops at size when every value of the list is less than the value
that the caller looks for. That is not an index of the list. The caller then
read memory after the end of the list. An assertion held the value inside the
range, but an assertion goes away in a build with NDEBUG. A build with NDEBUG
read data[3] of a list with three elements.

The result now stays below the size. Thus it is always an index that the
caller can use.

2. The decomposition read and wrote after the end of the peak buffer and the
valley buffer for a short signal.

The buffers hold one element for each sample. The code puts one point before
the peaks and one point after them, thus it needs room for two points more
than the number of peaks. A signal with fewer than three samples holds no
peak, and the code still wrote both points. For a signal with one sample
AddressSanitizer reported a heap-buffer-overflow. The function also read
x[1], which such a signal does not hold.

Such a signal holds nothing to take out. The function now gives an intrinsic
mode function with the value zero, and it reports the status 0. The
reconstruction of the signal stays correct, because the residue keeps the
whole signal. The new macro EMD_MINIMUM_SIZE holds the limit.

3. The decomposition built its envelopes from memory that held no value.

A signal that only rises or only falls holds no peak and no valley. The
detection then writes nothing into the buffers, but the code read the first
element of each buffer. That memory comes from malloc, thus it holds no
value. The envelopes now take the largest sample and the smallest sample of
the signal.

The static allocation of a vector asked for a pointer to uint32_t, but a
vector holds float values. A caller with a float buffer broke the rule about
the type of a pointer, and the unit test did exactly that. The parameter is
now a pointer to float, as in every other module of the library.

The tests cover a value on each side of the range of the list, a list with
one element, a signal with one sample and with two samples, and a signal that
only rises. A new property examines that the index of the search stays inside
the list for any value.

+++ tests/mocks/vector/Mock_vector.h
@@ -3,7 +3,7 @@
 
 #ifdef TEST
 vector_t vector_alloc(uint32_t size);
-vector_t vector_static_alloc(uint32_t size, uint32_t* mempool);
+vector_t vector_static_alloc(uint32_t size, float* mempool);
 void vector_add_point_at_index(vector_t* vector, uint32_t index, float data);
 void vector_add_from_array(vector_t* vector, uint32_t size, float* data);
 void vector_printf(vector_t* vector, print_t func);
