+++ vector/vector.h
@@ -18,7 +18,7 @@ typedef struct{
 }vector_t;
 
 vector_t vector_alloc(uint32_t size);
-vector_t vector_static_alloc(uint32_t size, uint32_t* mempool);
+vector_t vector_static_alloc(uint32_t size, float* mempool);
 void vector_add_point_at_index(vector_t* vector, uint32_t index, float data);
 void vector_add_from_array(vector_t* vector, uint32_t size, float* data);
 void vector_printf(vector_t* vector, print_t func);
